fixed some bond energy issues causing unrealistic torsion angles

+++ src/prxteinmpnn/utils/coordinates.py
@@ -42,8 +42,8 @@ def apply_noise_to_coordinates(
   key, coord_key = jax.random.split(key)
 
   def add_noise(coords: StructureAtomicCoordinates) -> StructureAtomicCoordinates:
-    noise = jax.random.normal(coord_key, coords.shape)
-    return coords + backbone_noise * noise
+    noise = jax.random.normal(coord_key, coords.shape, dtype=coords.dtype)
+    return (coords + backbone_noise * noise).astype(coords.dtype)
 
   def no_noise(coords: StructureAtomicCoordinates) -> StructureAtomicCoordinates:
     return coords
@@ -54,7 +54,13 @@ def apply_noise_to_coordinates(
     no_noise,
     coordinates,
   )
-  return noisy_coordinates, key
+  # Ensure output type matches input type explicitly if needed, 
+  # but the issue is likely that add_noise produces f64 while coordinates is f32.
+  # We should cast the result of cond or ensure add_noise respects input dtype.
+  # However, JAX random.normal might produce f64 if enable_x64 is True.
+  # Let's cast the noise to the coordinate dtype.
+  
+  return noisy_coordinates.astype(coordinates.dtype), key
 
 
 @jax.jit
